refactor: FirebaseBusinessException 디렉토리 이동
