+++ src/main/java/com/kustacks/kuring/message/firebase/FirebaseService.java
@@ -30,7 +30,7 @@ import java.util.stream.Collectors;
 public class FirebaseService {
 
     private static final String NOTIFICATION_TITLE = "새로운 공지가 왔어요!";
-    private final String DEV_SUFFIX = "dev";
+    private static final String DEV_SUFFIX = "dev";
     private final FirebaseMessaging firebaseMessaging;
     private final ObjectMapper objectMapper;
