+++ src/main/java/com/kustacks/kuring/message/firebase/exception/FirebaseSubscribeException.java
@@ -1,7 +1,6 @@
 package com.kustacks.kuring.message.firebase.exception;
 
 import com.kustacks.kuring.common.exception.code.ErrorCode;
-import com.kustacks.kuring.message.firebase.FirebaseBusinessException;
 
 public class FirebaseSubscribeException extends FirebaseBusinessException {
