+++ src/main/java/com/kustacks/kuring/message/firebase/exception/handler/FirebaseExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.kustacks.kuring.message.firebase.exception.handler;
 
 import com.kustacks.kuring.common.dto.ErrorResponse;
-import com.kustacks.kuring.message.firebase.FirebaseBusinessException;
+import com.kustacks.kuring.message.firebase.exception.FirebaseBusinessException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 public class FirebaseExceptionHandler {
 
     @ExceptionHandler
-    public ResponseEntity<ErrorResponse> FirebaseExceptionHandler(FirebaseBusinessException exception) {
+    public ResponseEntity<ErrorResponse> firebaseCommonExceptionHandler(FirebaseBusinessException exception) {
         log.error("[FirebaseBusinessException] {}", exception.getMessage());
 
         return ResponseEntity.status(exception.getErrorCode().getHttpStatus())
